+++ CHANGELOG.md
@@ -53,6 +53,8 @@
 - GP2-315 - route to market backend integration
 
 ### Implemented enhancements
+
+- GP2-449 - Extended video model for transcript field
 - GP2-582 - Product search handling multiple item selection
 - GP2-581 & GP2-586 - Product search spinner, 'no results message' and scroll area
 - GP2-528 - Horizontal rule styling updates